tank/requirements.txt
Jetson.GPIO
keyboard
pyPS4Controller